requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.7
Werkzeug==2.3.7
click==8.1.7 